Rename libpyastrochemtools to tools
