+++ src/converttolegacy.py.in
@@ -20,10 +20,10 @@
 #  along with Astrochem.  If not, see <http://www.gnu.org/licenses/>.
 
 import sys
-sys.path.append("@abs_top_srcdir@/python/libastrochemtools") #This line (first containing sys.path.append) will be removed at instalation.
+sys.path.append("@abs_top_srcdir@/python/") #This line (first containing sys.path.append) will be removed at instalation.
 sys.path.append("@pythondir@")
 import getopt
-from libastrochemtools import converttolegacy, listspecies
+from astrochem.tools import converttolegacy, listspecies
 
 PACKAGE_VERSION = "@PACKAGE_VERSION@"
 PACKAGE_BUGREPORT = "@PACKAGE_BUGREPORT@"
