+++ src/plabun.py.in
@@ -20,13 +20,13 @@
 #  along with Astrochem.  If not, see <http://www.gnu.org/licenses/>.
 
 import sys
-sys.path.append("@abs_top_srcdir@/python/libastrochemtools") #This line (first containing sys.path.append) will be removed at instalation.
+sys.path.append("@abs_top_srcdir@/python/") #This line (first containing sys.path.append) will be removed at instalation.
 sys.path.append("@pythondir@")
 import os.path
 import getopt
 import biggles
 from numpy import *
-from libastrochemtools import readabun,readabunlegacy, _totex_species, listspecies
+from astrochem.tools import readabun,readabunlegacy, _totex_species, listspecies
 
 PACKAGE_VERSION = "@PACKAGE_VERSION@"
 PACKAGE_BUGREPORT = "@PACKAGE_BUGREPORT@"
