+++ src/plroute.py.in
@@ -21,13 +21,13 @@
 #  along with Astrochem.  If not, see <http://www.gnu.org/licenses/>.
 
 import sys
-sys.path.append("@abs_top_srcdir@/python/libastrochemtools") #This line (first containing sys.path.append) will be removed at instalation.
+sys.path.append("@abs_top_srcdir@/python/") #This line (first containing sys.path.append) will be removed at instalation.
 sys.path.append("@pythondir@")
 import os.path
 import getopt
 import biggles
 from numpy import *
-from libastrochemtools import network_reader, readroutlegacy, readrout, _totex_species, listspecies, readfilesattrs
+from astrochem.tools import network_reader, readroutlegacy, readrout, _totex_species, listspecies, readfilesattrs
 
 PACKAGE_VERSION = "@PACKAGE_VERSION@"
 PACKAGE_BUGREPORT = "@PACKAGE_BUGREPORT@"
